requirements.txt
streamlit
openai
langchain
snowflake-connector-python
snowflake-snowpark-python
python-dotenv 
